End of Keymap fixes
