+++ src/auto/versions.txt
@@ -1,2 +1,2 @@
-Generated by gir (https://github.com/gtk-rs/gir @ 9e3cb65)
+Generated by gir (https://github.com/gtk-rs/gir @ 6fc5982)
 from gir-files (https://github.com/gtk-rs/gir-files @ f7e3c51)
